Cleanup CssClasses hierarchy, moved to shared
Moved org.eclipse.scout.rt.client.ui.CssClasses to
org.eclipse.scout.rt.shared.CssClasses

338668
